src/requirements.txt
Django
celery[librabbitmq,redis,auth,msgpack]
requests
qiniu
pyquery
redis
Markdown
PyMySQL
Twisted
djangorestframework